checks to ensure no file over-writing occurs

+++ src/new/activate_new.rs
@@ -24,47 +24,52 @@ end # module Test
 
 // TODO! Change this function to handle setting up files in a target directory
 fn create_default_files_for_env() -> std::io::Result<()> {
-    // let mut env_path = path::PathBuf::new();
-    // env_path.push(env_name);
+    // Check to ensure that you're not going to over-write any pre-existing user content...
+    let src_main_file = path::PathBuf::from("./src/Main.jl");
+    if src_main_file.exists() {
+        println!(
+            "\nActivated environment in {:?}.\nYou're ready to add packages!",
+            env::current_dir().expect("couldn't retrieve current directory")
+        );
+    } else {
+        let current_dir = env::current_dir()?;
 
-    // if env_path == current_dir()? {
-    //     let current_dir = env::current_dir()?;
-    // } else {
-    //     let current_dir = env_path;
-    // }
+        // Create ./src directory
+        let mut src_path = path::PathBuf::new();
+        src_path.push(&current_dir);
+        src_path.push("src/");
 
-    let current_dir = env::current_dir()?;
+        DirBuilder::new()
+            .recursive(true)
+            .create(src_path)
+            .expect("Could not create `/src/` directory");
 
-    // Create ./src directory
-    let mut src_path = path::PathBuf::new();
-    src_path.push(&current_dir);
-    src_path.push("src/");
+        // write `Main.jl` to current_dir/src/Main.jl
+        let mut jl_main_file =
+            File::create("./src/Main.jl").expect("could not create ./src/Main.jl file");
 
-    DirBuilder::new()
-        .recursive(true)
-        .create(src_path)
-        .expect("Could not create `/src/` directory");
+        // write JL_SCRIPT_CONTENTS to the julia main file
+        write!(jl_main_file, "{}", JL_SCRIPT_CONTENTS)?;
 
-    // write `Main.jl` to current_dir/src/Main.jl
-    let mut jl_main_file =
-        File::create("./src/Main.jl").expect("could not create ./src/Main.jl file");
+        let mut tests_path = path::PathBuf::new();
+        tests_path.push(&current_dir);
+        tests_path.push("./tests/");
 
-    // write JL_SCRIPT_CONTENTS to the julia main file
-    write!(jl_main_file, "{}", JL_SCRIPT_CONTENTS)?;
+        DirBuilder::new()
+            .recursive(true)
+            .create(tests_path)
+            .expect("Could not create `/tests/` directory");
 
-    let mut tests_path = path::PathBuf::new();
-    tests_path.push(&current_dir);
-    tests_path.push("./tests/");
+        let mut jl_runtests_file = File::create("./tests/run_tests.jl")
+            .expect("could not create `./tests/run_tests.jl` file");
 
-    DirBuilder::new()
-        .recursive(true)
-        .create(tests_path)
-        .expect("Could not create `/tests/` directory");
+        write!(jl_runtests_file, "{}", JL_RUNTESTS_CONTENTS)?;
 
-    let mut jl_runtests_file =
-        File::create("./tests/run_tests.jl").expect("could not create `./tests/run_tests.jl` file");
-
-    write!(jl_runtests_file, "{}", JL_RUNTESTS_CONTENTS)?;
+        println!(
+            "\nActivated new project environment and files in {:?}.\nYou're ready to add packages!",
+            env::current_dir().expect("couldn't retrieve current directory")
+        );
+    }
 
     Ok(())
 }
@@ -110,13 +115,9 @@ pub fn activate_env_in_current_dir(julia: &mut Julia) {
         .expect("Result is an error");
 
     create_default_files_for_env().expect("Couldn't write default files for the environment");
-
-    println!(
-        "\nActivated new project environment in {:?}.\nYou're ready to add packages!",
-        env::current_dir().expect("couldn't retrieve current directory")
-    );
 }
 
+// TODO! create default files unless ./src/Main.jl & /tests/run_tests.jl files exist
 pub fn activate_env_w_name(julia: &mut Julia, env_name: &str) {
     println!("\nActivating environment \"{}\"\n", &env_name);
